+++ test/ModuleBundler.js
@@ -3,7 +3,7 @@ import path from 'path';
 import Yazl from 'yazl';
 import ModuleBundler from '../src/lib/ModuleBundler';
 
-describe('ModuleBundler', async function () {
+describe('ModuleBundler', function () {
   this.timeout(5000);
 
   const servicePath = path.resolve(__dirname, './1.0');
@@ -13,7 +13,7 @@ describe('ModuleBundler', async function () {
     servicePath,
   }, artifact);
 
-  const { dependencies } = require(`${servicePath}/package.json`);
+  const { dependencies } = require(`${servicePath}/package.json`); // eslint-disable-line
 
   before(async () => {
     await moduleBundler.bundle({});
@@ -22,7 +22,7 @@ describe('ModuleBundler', async function () {
   for (const dep in dependencies) {
     it(`Has bundled dependency ${dep}`, async () => {
       assert(
-        moduleBundler.modules.some(({ name }) => name === dep)
+        moduleBundler.modules.some(({ name }) => name === dep),
       );
     });
   }