+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.5.1
+- Added coloring and config output through logging
+- Changed deploy hooks, single function fixed
+- By default `uglify` is off
+  - `babeli` is needed in future
+
 ## 0.5.0.alpha.2
 - Added `deepExclude` module checks when bundling nested module `node_modules` directories
 ## 0.3.0 -> 0.4.0